requirements.txt
streamlit
google-generativeai
beautifulsoup4
python-dotenv
fpdf
PyPDF2